requirements.txt
flask
pymessenger
flask-sqlalchemy
python-dotenv
flask-migrate
mysqlclient
pandas
BeautifulSoup4
tabulate
psycopg2
gunicorn
matplotlib
ipychart
